+++ tb_rs/src/lib.rs
@@ -59,205 +59,13 @@ use std::{collections::HashMap, error::Error, ffi::CStr, os::raw::c_char};
 //     name: &'static str,
 // }
 
-mod verif {
-    use std::fmt;
+// includes verif/mod.rs
+mod verif;
 
-    #[derive(Clone)]
-    pub struct VirtualInterface;
-
-    pub trait Objectify {
-        fn get_name(&self) -> String;
-    }
-
-    pub trait PhasingA {
-        fn get_phase(&self) -> Phase;
-
-        /// Top-Down configuration
-        fn configure(&mut self);
-    }
-
-    pub trait PhasingB {
-        fn connect<T: PhasingB>(&self, other: T);
-        fn elaborate(&self);
-        /**
-         TODO: need concept of "time" in order to distinguish between
-         non-time-consuming and time-consuming fn's
-        */
-        /*async*/
-        fn run(&self);
-        fn shutdown(&self);
-    }
-
-    #[derive(Clone, Debug)]
-    pub enum Phase {
-        Allocated,
-        Configured,
-        Connected,
-        Run,
-    }
-
-    // println! for Phase
-    impl fmt::Display for Phase {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            match self {
-                Phase::Allocated => write!(f, "Allocated"),
-                Phase::Configured => write!(f, "Configured"),
-                Phase::Connected => write!(f, "Connected"),
-                Phase::Run => write!(f, "Run"),
-            }
-        }
-    }
-
-    // pub enum Components {
-    //     Environment(Object),
-    //     Agent,
-    //     Monitor,
-    //     Driver,
-    //     Sequencer,
-    // }
-
-    // // this is like a C function-pointer
-    // pub enum Object {
-    //     Component(u32),
-    //     Transaction,
-    // }
-
-    // impl Object {
-    //     fn unref(&self) -> Components {
-    //         match self {
-    //             Object::Component(i) => Components::Environment(*i),
-    //             _ => Components::Agent,
-    //         }
-    //     }
-    // }
-
-    // pub struct Component<PHASE, OBJ> {
-    //     phase: PHASE,
-    //     object: OBJ,
-    // }
-
-    // pub fn ComponentBuilder<OBJ>(obj: OBJ) -> Component<Phase, Components> {
-    //     let a = Object::Component(1);
-
-    //     Component {
-    //         phase: Phase::Allocated,
-    //         object: a.unref(),
-    //     }
-    // }
-}
-
-mod my_agent {
-    use crate::verif::{self, Objectify, PhasingA, PhasingB, VirtualInterface};
-    use std::collections::HashMap;
-    pub struct Agent {
-        pub name: &'static str,
-        // heterogenous hashmap of Trait Objects
-        pub component_db: HashMap<String, Box<dyn PhasingA>>,
-        pub phase: verif::Phase,
-        pub is_active: bool,
-    }
-
-    pub struct Monitor {
-        pub name: &'static str,
-        // heterogenous hashmap of Trait Objects
-        pub component_db: HashMap<String, Box<dyn PhasingA>>,
-        pub phase: verif::Phase,
-        pub vif: verif::VirtualInterface,
-    }
-
-    impl<'a> Objectify for Monitor {
-        fn get_name(&self) -> String {
-            self.name.to_string()
-        }
-    }
-
-    impl<'a> PhasingA for Monitor {
-        fn get_phase(&self) -> verif::Phase {
-            dbg!("{}", &self.phase);
-            println!("{}: {}", self.get_name(), &self.phase);
-            self.phase.clone()
-        }
-
-        fn configure(&mut self) {
-            // Top-Down configuration
-            self.component_db.values_mut().for_each(|v| {
-                v.configure();
-            });
-        }
-    }
-
-    pub struct Driver {
-        pub name: &'static str,
-        // heterogenous hashmap of Trait Objects
-        pub component_db: HashMap<String, &'static mut dyn PhasingA>,
-        pub phase: verif::Phase,
-        pub vif: verif::VirtualInterface,
-    }
-
-    impl Objectify for Driver {
-        fn get_name(&self) -> String {
-            self.name.to_string()
-        }
-    }
-
-    impl PhasingA for Driver {
-        fn get_phase(&self) -> verif::Phase {
-            dbg!("{}", &self.phase);
-            println!("{}: {}", self.get_name(), &self.phase);
-            self.phase.clone()
-        }
-
-        fn configure(&mut self) {
-            // Top-Down configuration
-            for v in self.component_db.values_mut() {
-                v.configure();
-            }
-        }
-    }
-
-    impl Objectify for Agent {
-        fn get_name(&self) -> String {
-            self.name.to_string()
-        }
-    }
-
-    impl PhasingA for Agent {
-        fn get_phase(&self) -> verif::Phase {
-            dbg!("{}", &self.phase);
-            println!("{}: {}", self.get_name(), &self.phase);
-            self.phase.clone()
-        }
-
-        fn configure(&mut self) {
-            // allocate the child monitor and driver
-
-            self.component_db.insert(
-                String::from("kjsdkfjk"),
-                Box::new(Monitor {
-                    name: "sldkfj",
-                    component_db: HashMap::new(),
-                    phase: self.phase.clone(),
-                    vif: VirtualInterface {},
-                }),
-            );
-
-            if self.is_active {
-                let drvr = Driver {
-                    name: "jklwej",
-                    component_db: HashMap::new(),
-                    phase: self.phase.clone(),
-                    vif: VirtualInterface {},
-                };
-                self.component_db.insert(drvr.get_name(), Box::new(drvr));
-            }
-
-            // Top-Down configuration
-            for v in self.component_db.values_mut() {
-                v.configure();
-            }
-        }
-    }
-}
+/**
+  example of a standalone interface Agent
+*/
+mod my_agent;
 
 mod my_env {
